requirements.txt
aiohttp>=3.9
click>=8.1
PyYAML>=6.0
python-dotenv>=1.0
py-clob-client>=1.9
tabulate>=0.9
